// src/app/shared/models/form-data.ts
export class FormData {
    firstName: string = '';
    lastName: string = '';
    address: string = '';
    city: string = '';
    country: string = '';
    code: string = '';

    clear() {
        this.firstName = '';
        this.lastName = '';
        this.address = '';
        this.city = '';
        this.country = '';
        this.code = '';
    }
}


export class Shipping {
    firstName: string = '';
    lastName: string = '';
    address: string = '';
    city: string = '';
    country: string = '';
    code: string = '';
}